Add ability to customize amount of hops for IMDSv2
This allows implementors to fix docker containers on EC2 not being able
to get a session token.

With docker containers on a EC2 host, the docker container could not get
the session token. By increasing the hop limit they can allow the request
to hop from docker to the host and then to AWS.